# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Task Manager API
  description: REST API for the Task Manager application
  version: "1.0.0"
servers:
  - url: http://localdev.taskmanager.com/api/v1
    description: Development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /tasks:
    get:
      tags: [Tasks]
      summary: List user tasks
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: priority
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: sort
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: from
          schema:
            type: string
        - in: query
          name: to
          schema:
            type: string
      responses:
        "200":
          description: Task list
    post:
      tags: [Tasks]
      summary: Create a task
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                description:
                  type: string
                priority:
                  type: string
                due_date:
                  type: string
      responses:
        "201":
          description: Task created
        "422":
          description: Validation failed
...